Moved keys cmd to top level
