+++ cmd/utils.go
@@ -6,8 +6,8 @@ import (
 	"github.com/bgentry/speakeasy"
 	"github.com/pkg/errors"
 	"github.com/spf13/viper"
-	data "github.com/tendermint/go-data"
 	keys "github.com/tendermint/go-crypto/keys"
+	data "github.com/tendermint/go-wire/data"
 )
 
 const PassLength = 10
