+++ glide.yaml
@@ -13,12 +13,22 @@ import:
   version: unstable
 - package: github.com/tendermint/go-wire
   version: unstable
+  subpackages:
+  - data
+  - data/base58
 - package: golang.org/x/crypto
   subpackages:
   - blowfish
   - nacl/secretbox
   - openpgp/armor
   - ripemd160
+- package: github.com/bgentry/speakeasy
+- package: github.com/gorilla/handlers
+- package: github.com/gorilla/mux
+- package: github.com/pkg/errors
+- package: github.com/spf13/cobra
+- package: github.com/spf13/viper
+- package: gopkg.in/go-playground/validator.v9
 testImport:
 - package: github.com/mndrix/btcutil
 - package: github.com/stretchr/testify
